deps: update superchain registry
Updates the version in the `go.mod` to https://github.com/ethereum-optimism/superchain-registry/commit/7b48d381868687d7a2124dc092434b34d4e0189c
This was done with the following command:

```bash
$ go get -v github.com/ethereum-optimism/superchain-registry/superchain
```

This commit includes the latest deployments of the contract
implementations. Pulling in this new version is required to
build the spell for the `superchain-ops` repo.