contracts-bedrock: modernize devnet artifacts
The internal-devnet was deployed using the hh deploy scripts,
meaning that the names of the files are different. It also
has dead artifacts such as the `SystemDictator` and the
`PortalSender`. This migrates the names of the files to
the modern style as well as deletes artifacts that are not
needed.