op-upgrade: cleanup implementation
Use a higher level API that is more appropriate for building
calldata for calls. The lower level API did not append the 4byte
selectors and that had to be done manually. Using the low level
API was error prone so it should be avoided. The high level API
does exactly what we want it to do - ABI encode the data and then
append it to the correct 4byte selector.

To review this PR, ensure that the arguments to the old `PackValues`
functions are the same as the new `Pack` function, excluding the
first argument which is the name of the function to ABI encode
for.