ci-builder: include solc 0.8.19
The other 2 compiler versions are included in `ci-builder` so we should
also include `0.8.19`. This will save us from needing to download the
compiler in CI many times.

The following lines are logged in CI:

```
3[2K[⠆]3[2Ktalling solc version 0.8.19[⠰]3[2Ktalling solc version 0.8.19[⠔] Installing solc version 0.8.19
3[2K[⠒]3[2Kcessfully installed solc 0.8.19[⠑] Successfully installed solc 0.8.19
```

These lines are not logged for the other 2 solc versions that are
included in `ci-builder` because they are already downloaded.