op-node: Hide rethDB option
If the rethDB option is used without the proper build configuration,
the op-node will panic. Note that we do not have this build configuration
in our releases. This hides the option to use it, but does not fully
remove it. There was some duplication required to easily hide the
command line flag while still being able to use it for testing purposes.

This commit should be reverted when rethDB is stable.