ci: solc building requires xlarge
Modularize out a small change from https://github.com/ethereum-optimism/optimism/pull/7928
where any `forge build` runs oom unless at least an `xlarge` box is used
in circleci. As we have added more contracts to the repo, building the
contracts has consumed more and more CPU/memory. The deploy script is
quite large and uses a lot of memory, so adding it to our tests makes
the tests use more memory. I have not noticed the tests taking a longer
time to run locally.

We could also do a better job at encapsulating where solc builds happen
such that it only happens once and the results are cached as much as
possible. Making the boxes bigger in the meantime works, as it should
also speed up CI for the tasks where the boxes are made bigger.