contracts-bedrock: cleanup expect emits
Modern foundry added the ability to `vm.expectEmit(address)`. This is an
alias for `vm.expectEmit(bool,bool,bool,bool,address)` where each of the
booleans are true. It is important to include the address of the
contract this is expecting to emit the event. If all of the bools are
going to be true, might as well use the short hand to simplify the
source code as well as the bytecode.