contracts-bedrock: migrate to new storage snapshot
Delete the old storage layout now that https://github.com/ethereum-optimism/optimism/pull/8243
has been merged which adds another snapshots concept of both the abi and
the storage layout.

The old method of keeping the storage layout lockfile around is now
deprecated. Once we delete it, we will no longer be able to track
changes from old versions of the contracts as easily. We can reintroduce
the shell script to a future commit if necessary.