monorepo: make devnet-up just works
Ensure that the submodules exist as part of the
`make devnet-up` command. This enables somebody
to clone the repo and run a single command to bring
up the entire devnet without needing to worry about
running any other commands. Also replace some calls
in the CI to use `make` instead of `git` directly
for when installing the submodules.