op-chain-ops: cleanup immutables
We really only need to know if a contract has immutables and not what
the immmutables are to write good tests to ensure that all contracts
that have immutables are covered in the L2 genesis generation. Adding in
the actual immutables references as a string resulted in a lot of diffs
because the ast ids where used in them. This updates the bindings to use
a boolean instead of the string. Now we can still tell if a contract has
immutables and we dont care about what the offsets are. This will not
result in a large diff for the smart contracts when ast ids change.
It will only flip a bool when immutables are added or removed from a
contract.