op-e2e: Ignore i/o timeouts in receipts fetching
Ignore i/o timeouts when fetching receipts. This is causing spurious failures in the indexer tests. We already have a 2 minute timeout within this method via an internal context, so there is no value in erroring out on what could (potentially) be an unrelated network failure.

If there is some other underlying error in the indexer tests, we'll be able to see that because the tests will fail with `context deadline exceeded`.