op-node/flags: Add easy way to deprecate flags
This adds a deprecated flags section to warn users that the flag will be removed
in a future version of the software. I had to modify the P2P Flags structure to
make it work nicely.

I also added a test to ensure that deprecated flags are hidden & that all flags
have at least one env var.