ci: Slim down CI builder (#9339)
* ci: Slim down CI builder

The CI builder image is downloaded many hundreds of times per day. The larger it is, the longer it takes for each CI executor to spin up which increases overall CI runtime. It also causes a significant amount of bandwidth costs from GCP Artifact Registry. This PR reduces the size of CI builder by moving the Rust toolchain to a dedicated builder image which will only be used with the jobs that require it. The Rust toolchain accounts for more than 50% of the size of the builder image.

* add binutils

* fix path

* Update SVM

* have to source bash first

* bad copy

* add back elf tools