fix(ct): bugs in DelayedWETH (#9899)
* fix(ct): bugs in DelayedWETH

Fixes two bugs in Delayed WETH. First was a limitation of the
recover function that would have made it relatively easy to DoS
the function by simply withdrawing small amounts of ETH every time
the owner attempted to withdraw the full balance. Second is a bug
in hold function where the approval was the wrong way around.

* Update Kontrol snapshot

---------

Co-authored-by: inphi <mlaw2501@gmail.com>