op-upgrade: delete dead version (#10013)
* op-upgrade: delete dead version

Deletes the `op-upgrade-mcp` binary in favor of the `op-upgrade` binary.
This enables us to leverage the `superchain-regsitry` without needing to
hardcode addresses in one off locations.

`op-upgrade-mcp` existed because `superchain-regsitry` assumes that
there are implementations that can be shared between networks, ie
post MCP world, meaning there are no immutables in the code that
are network specific.

* op-upgrade: apply superchain config to L1 upgrade work, e.g. to source superchainConfig.Config.SuperchainConfigAddr for L1CrossDomainMessenger

---------

Co-authored-by: protolambda <proto@protolambda.com>