op-proposer: remove dep on op-bindings (#10218)
Prepares the `op-proposer` to no longer depend
on the `op-bindings` package. This will enable
releases of the `op-proposer` to work with specific
releases of the contracts more easily by using `abigen`
at a specific release of the contracts to create bindings.

We will be gutting `op-bindings` as a package because
it is not versioned, implying all versions must be backwards
and forwards compatible. It also adds a lot of overhead to
smart contract development with having to create the bindings
and commit in a diff every time there is a solidity source file
change.